requirements.txt
aspose-words==22.7.0
camelot-py==0.10.1
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
distro==1.7.0
et-xmlfile==1.1.0
fire==0.4.0
fonttools==4.34.4
ghostscript==0.7
loguru==0.6.0
lxml==4.9.1
numpy==1.23.1
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.4.3
pdf2docx==0.5.5
pdfminer.six==20220524
pdftopng==0.2.3
pycparser==2.21
PyMuPDF==1.20.1
PyPDF2==2.8.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
six==1.16.0
tabula-py==2.4.0
tabulate==0.8.10
termcolor==1.1.0
typing-extensions==4.3.0
